requirements.txt
Flask==2.1.1
Flask_Login==0.6.0
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_Table==0.5.0
Flask_WTF==1.0.1
nltk==3.7
pandas==1.4.3
Pillow==9.3.0
preprocessor==1.1.3
requests==2.28.1
tweet_preprocessor==0.6.0
